+++ docs/source/isaac_teleop.mdx
@@ -262,7 +262,7 @@ commanded only while the clutch is engaged.
 
 **Customizing the reset pose.** The reset pose ships as a built-in default (a comfortable mid-range
 pose) and works out of the box — you do **not** need to record anything. To tailor it to your setup,
-back-drive the arm to the pose you want and run `python record_reset_pose.py` (in the example
+back-drive the arm to the pose you want and run `python override_reset_pose.py` (in the example
 folder); it writes the current joints to `reset_pose.json`, which then takes priority over the
 built-in default on the next run. That file is git-ignored (user-local), so your override stays on
 your machine, and both `teleoperate.py` and `record.py` honor it.
